pyproject.toml
[project]
name = "ecomod-hydro"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "bleak>=1.1.1",
    "gradio>=4.44.1",
    "huesdk>=1.8",
    "python-kasa>=0.7.7",
]
